+++ NEWS.md
@@ -3,8 +3,8 @@ Development
 
 ### Features
 * New loading button styles (#12132)
+* [WIP] Export/import organization/user metadata to allow user migration (#12271, #12304)
 * New force param in EUMAPI organization users destroy operation to force deletion even with unregistered tables (#11654).
-* [WIP] Export/import organization/user metadata to allow user migration (#12271)
 * Removed the usage of the `organizations_admin` feature flag (#12131)
 
 ### Bug fixes / enhancements